+++ src/app/map-actions/map-actions.component.ts
@@ -21,15 +21,16 @@ export class MapActionsComponent implements OnInit {
 	private MapService: MapService;
 	private configSettings: Config;
 	public baselayers = [] as any;
+	private currentBaseLayer: any;
 	public workflowLayers;
 	public overlayLayers;
 	public currentZoom: number = 4;
 	public selectedWorkflow: any;
 
-  	constructor(private _mapService: MapService, private _configService: ConfigService, private _workflowService: WorkflowService) {
+	constructor(private _mapService: MapService, private _configService: ConfigService, private _workflowService: WorkflowService) {
 		this.MapService = _mapService;
 		this.configSettings = this._configService.getConfiguration();
-	  }
+	}
 
 	ngOnInit(): void {
 		
@@ -53,7 +54,10 @@ export class MapActionsComponent implements OnInit {
 		this._workflowService.selectedWorkflow.subscribe(res => {
 			this.selectedWorkflow = res;
 		})
-
+		// Get current basemap
+		this.MapService.baseLayer.subscribe((currentBaseLayer: any) => {
+			this.currentBaseLayer = currentBaseLayer;
+		});
 	}
 
 	public SetBaselayer(LayerName: string) {