+++ CHANGELOG.md
@@ -27,9 +27,34 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed  
 
 - 
+
 ### Security  
 
 - 
+
+## [v1.4.0-beta](https://github.com/USGS-WiM/StreamStats-National/releases/tag/v1.4.0-beta) - 2023-03-15
+
+### Added 
+
+- "WFIGS - Wildland Fire Perimeters Full History" layer to fire hydrology workflows
+- "Interagency Fire Perimeter History - All Years" layer to fire hydrology workflows
+
+### Changed  
+
+- The Report tab now automatically opens after completing a workflow
+- Fire map services due to NIFC updates
+- Unavailable grid query service values are evaluated as null rather than -9999
+- Improved notification when dependent services are unavailable
+
+### Removed 
+
+- "2021 Fire Perimeters" layer from fire hydrology workflows
+
+### Fixed  
+
+- In "Delineation" workflow, only return Basin Characteristics that are available at the clicked point
+- In "Query by Fire Perimeter", fixed bug causes it to not select a perimeter
+- Query strings used to query fire perimeter map layer geometry in queryBurnedArea function 
 ## [v1.3.0-beta](https://github.com/USGS-WiM/StreamStats-National/releases/tag/v1.3.0-beta) - 2023-01-11
 
 ### Added